requirements.txt
flask
flasgger
groq
python-dotenv